Disable DigestUpdate->DigestInit transit
For mb_md5 (multi-buffer) provider, a transit from DigestUpdate
to DigestInit will cause the allocated resources in mb_md5 be
unreleased.

So, when a lot of such transits happen, resources in mb_md5 will
be exhausted. And any new DigestInit request cannot be fufilled.

As a workround, disable DigestUpdate->DigestInit transit when
mb_md5 is involved in the test.

Signed-off-by: Guodong Xu <guodong.xu@linaro.org>